Updated Ansible installation steps in Jenkinsfile.

+++ Jenkinsfile
@@ -8,8 +8,13 @@ pipeline{
 	    }
 	    stage('Installing Ansible') {
 	        steps {
-		    sh 'sudo apt-get update -y && sudo apt-get upgrade -y'
-		    sh 'sudo apt install -y wget tree unzip ansible python3-pip python3-apt'
+		    script {
+		        def ansible_exists = fileExists '/usr/bin/ansible'
+			if (ansible_exists) {
+			    echo "Skipping Ansible install -already exists"
+		        } else {
+		            sh 'sudo apt-get update -y && sudo apt-get upgrade -y'
+		            sh 'sudo apt install -y wget tree unzip ansible python3-pip python3-apt'
 		    }
 		}
 		stage('Third Stage') {
